// api/models/NewResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class NewResponse {
    'taskName': string;
    'initialScores'?: any;
    'convHistory'?: any;
    'finalScores'?: any;
    'userId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "taskName",
            "baseName": "task_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "initialScores",
            "baseName": "initial_scores",
            "type": "any",
            "format": ""
        },
        {
            "name": "convHistory",
            "baseName": "conv_history",
            "type": "any",
            "format": ""
        },
        {
            "name": "finalScores",
            "baseName": "final_scores",
            "type": "any",
            "format": ""
        },
        {
            "name": "userId",
            "baseName": "user_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return NewResponse.attributeTypeMap;
    }

    public constructor() {
    }
}